Support scanning of buildScripts / used gradle plugins
For this, the buildScriptProjects configuration option was introduced.
It respects all other configuration like 'configurations'.

The implementation is utilizing the existing ProjectReader for this
because the required fields (`configurations` and `dependencies`) can
also be found in ScriptHandler. There is only a change required: to use
a general data structure for accessing these fields.